better future proof
Reworked code to no longer rely on reference cues, but instead build individual scripts for each output desired. I think in the long run this is going to be better and more reusable than having a script that turned off multiple projects, then you would have to redo it for every iteration of projector groupings.