reset test script
This script used to help reset the test file when building create slideshow updated script. It resets the properties of all video cues in the current cue list and selects them to decrease time in between tests. Find way to define list to set selected for greater reusability, modifiyng all video cues is not transportable.